+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated dependencies to latest
   - added theme files to stylelint ignore
   - fixed other stylelint errors
+- Added functionality to favorite a song ([fixes #323](https://github.com/Mastermindzh/tidal-hifi/issues/323))
+  - Added a hotkey to favorite ("Add to collection") songs: Control+a
+  - Added the "favorite" field in the `mediaInfo` and the API `/current` endpoint
+  - Added an endpoint to toggle favoriting a song: `http://localhost:47836/favorite/toggle`
 
 ## [5.7.1]